# api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Boost rest APi",
    "version": "2.0",
    "title": "boost rest api",
    "contact": {
      "name": "Ats dev team"
    },
    "license": {}
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/apiAfin/courses": {
      "get": {
        "tags": [
          "course-resource"
        ],
        "summary": "find all courses by page",
        "operationId": "getCoursesPaginateUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "size",
            "description": "size",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "in": "body",
            "name": "page",
            "description": "page",
            "required": false,
            "schema": {
              "type": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«Course»"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/courses/": {
      "post": {
        "tags": [
          "course-resource"
        ],
        "summary": "create new course",
        "operationId": "createModuleUsingPOST_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "course",
            "description": "course",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "course-resource"
        ],
        "summary": "update course",
        "operationId": "updateModuleUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "course",
            "description": "course",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Course"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/courses/all": {
      "get": {
        "tags": [
          "course-resource"
        ],
        "summary": "find all courses",
        "operationId": "getAllCoursesUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Course"
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/courses/{id}": {
      "get": {
        "tags": [
          "course-resource"
        ],
        "summary": "find course by id",
        "operationId": "getCourseUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/apiAfin/courses/{id}/module": {
      "post": {
        "tags": [
          "course-resource"
        ],
        "summary": "add module to course the id param is for course",
        "operationId": "addModuleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "courseModule",
            "description": "courseModule",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseModule"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Course"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/homeworks": {
      "get": {
        "tags": [
          "homework-resource"
        ],
        "summary": "find all homework by page",
        "operationId": "getCoursesPaginateUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "size",
            "description": "size",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "in": "body",
            "name": "page",
            "description": "page",
            "required": false,
            "schema": {
              "type": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«Homework»"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/homeworks/all": {
      "get": {
        "tags": [
          "homework-resource"
        ],
        "summary": "find all homework",
        "operationId": "getAllCoursesUsingGET_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Homework"
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/marks": {
      "get": {
        "tags": [
          "marks-resources"
        ],
        "summary": "find all marks by page",
        "operationId": "getCoursesPaginateUsingGET_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "size",
            "description": "size",
            "required": false,
            "schema": {
              "type": "int"
            }
          },
          {
            "in": "body",
            "name": "page",
            "description": "page",
            "required": false,
            "schema": {
              "type": "int"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Page«Marks»"
            }
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/modules/": {
      "post": {
        "tags": [
          "course-module-resource"
        ],
        "summary": "create new modules",
        "operationId": "createModuleUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "courseModule",
            "description": "courseModule",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseModule"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "course-module-resource"
        ],
        "summary": "update modules",
        "operationId": "updateModuleUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "courseModule",
            "description": "courseModule",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseModule"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CourseModule"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/modules/{id}/addmaterials": {
      "post": {
        "tags": [
          "course-module-resource"
        ],
        "summary": "create new modules",
        "operationId": "createModuleUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "moduleMaterial",
            "description": "moduleMaterial",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ModuleMaterial"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/apiAfin/questions/": {
      "post": {
        "tags": [
          "question-resources"
        ],
        "summary": "create new question",
        "operationId": "createModuleUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "question",
            "description": "question",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HwQuestion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "question-resources"
        ],
        "summary": "update question",
        "operationId": "updateQuestionsUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "hwQuestion",
            "description": "hwQuestion",
            "required": true,
            "schema": {
              "$ref": "#/definitions/HwQuestion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/HwQuestion"
            }
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ModuleMaterial": {
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "url": {
          "type": "string"
        }
      }
    },
...
    "Marks": {
      "properties": {
        "hwQuestion": {
          "$ref": "#/definitions/HwQuestion"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "LocalDate": {
      "properties": {
        "chronology": {
          "$ref": "#/definitions/IsoChronology"
        },
        "dayOfMonth": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "dayOfYear": {
          "type": "integer",
          "format": "int32"
        },
        "era": {
          "$ref": "#/definitions/Era"
        },
        "leapYear": {
          "type": "boolean"
        },
        "month": {
          "type": "integer",
          "format": "int32"
        },
        "monthValue": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ZoneOffset": {
      "properties": {
        "id": {
          "type": "string"
        },
        "rules": {
          "$ref": "#/definitions/ZoneRules"
        },
        "totalSeconds": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Solutions": {
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "registering": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "text": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
    "ZonedDateTime": {
      "properties": {
        "chronology": {
          "$ref": "#/definitions/Chronology"
        },
        "dayOfMonth": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "dayOfYear": {
          "type": "integer",
          "format": "int32"
        },
        "hour": {
          "type": "integer",
          "format": "int32"
        },
        "minute": {
          "type": "integer",
          "format": "int32"
        },
        "month": {
          "type": "string",
          "enum": [
            "JANUARY",
            "FEBRUARY",
            "MARCH",
            "APRIL",
            "MAY",
            "JUNE",
            "JULY",
            "AUGUST",
            "SEPTEMBER",
            "OCTOBER",
            "NOVEMBER",
            "DECEMBER"
          ]
        },
        "monthValue": {
          "type": "integer",
          "format": "int32"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "offset": {
          "$ref": "#/definitions/ZoneOffset"
        },
        "second": {
          "type": "integer",
          "format": "int32"
        },
        "year": {
          "type": "integer",
          "format": "int32"
        },
        "zone": {
          "$ref": "#/definitions/ZoneId"
        }
      }
    },
    "Page«Marks»": {
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Marks"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "CourseModule": {
      "properties": {
        "homework": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Homework"
          }
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "moduleMaterials": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ModuleMaterial"
          }
        },
        "opening": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "sommary": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "Sort": {},
    "Chronology": {
      "properties": {
        "calendarType": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
    "ZoneOffsetTransition": {
      "properties": {
        "dateTimeAfter": {
          "$ref": "#/definitions/LocalDateTime"
        },
        "dateTimeBefore": {
          "$ref": "#/definitions/LocalDateTime"
        },
        "duration": {
          "$ref": "#/definitions/Duration"
        },
        "gap": {
          "type": "boolean"
        },
        "instant": {
          "$ref": "#/definitions/Instant"
        },
        "overlap": {
          "type": "boolean"
        }
      }
    },
    "Course": {
      "properties": {
        "closing": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "courseModules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CourseModule"
          }
        },
        "discussions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Discussion"
          }
        },
        "domain": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "opening": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "registring": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "teachers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Teacher"
          }
        },
        "title": {
          "type": "string"
        }
      }
    },
    "ZoneOffsetTransitionRule": {
      "properties": {
        "dayOfMonthIndicator": {
          "type": "integer",
          "format": "int32"
        },
        "dayOfWeek": {
          "type": "string",
          "enum": [
            "MONDAY",
            "TUESDAY",
            "WEDNESDAY",
            "THURSDAY",
            "FRIDAY",
            "SATURDAY",
            "SUNDAY"
          ]
        },
        "localTime": {
          "$ref": "#/definitions/LocalTime"
        },
        "midnightEndOfDay": {
          "type": "boolean"
        },
        "month": {
          "type": "string",
          "enum": [
            "JANUARY",
            "FEBRUARY",
            "MARCH",
            "APRIL",
            "MAY",
            "JUNE",
            "JULY",
            "AUGUST",
            "SEPTEMBER",
            "OCTOBER",
            "NOVEMBER",
            "DECEMBER"
          ]
        },
        "timeDefinition": {
          "type": "string",
          "enum": [
            "UTC",
            "WALL",
            "STANDARD"
          ]
        }
      }
    },
...
    "HwQuestion": {
      "properties": {
        "answer": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "proposition1": {
          "type": "string"
        },
        "proposition2": {
          "type": "string"
        },
        "proposition3": {
          "type": "string"
        },
        "proposition4": {
          "type": "string"
        },
        "qmark": {
          "type": "number",
          "format": "double"
        },
        "statement": {
          "type": "string"
        }
      }
    },
    "ZoneId": {
      "properties": {
        "id": {
          "type": "string"
        },
        "rules": {
          "$ref": "#/definitions/ZoneRules"
        }
      }
    },
    "LocalTime": {
      "properties": {
        "hour": {
          "type": "integer",
          "format": "int32"
        },
        "minute": {
          "type": "integer",
          "format": "int32"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "second": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "TemporalUnit": {
      "properties": {
        "dateBased": {
          "type": "boolean"
        },
        "durationEstimated": {
          "type": "boolean"
        },
        "timeBased": {
          "type": "boolean"
        }
      }
    },
    "Instant": {
      "properties": {
        "epochSecond": {
          "type": "integer",
          "format": "int64"
        },
        "nano": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Teacher": {
      "properties": {
        "birthday": {
          "$ref": "#/definitions/LocalDate"
        },
        "contact": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "level": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "profession": {
          "type": "string"
        },
        "profile": {
          "type": "string"
        },
        "registring": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "surnames": {
          "type": "string"
        }
      }
    },
    "Duration": {
      "properties": {
        "nano": {
          "type": "integer",
          "format": "int32"
        },
        "negative": {
          "type": "boolean"
        },
        "seconds": {
          "type": "integer",
          "format": "int64"
        },
        "units": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TemporalUnit"
          }
        },
        "zero": {
          "type": "boolean"
        }
      }
    },
    "IsoChronology": {
      "properties": {
        "calendarType": {
          "type": "string"
        },
        "id": {
          "type": "string"
        }
      }
    },
...
    "Era": {
      "properties": {
        "value": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Page«Homework»": {
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Homework"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "Page«Course»": {
      "properties": {
        "content": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Course"
          }
        },
        "first": {
          "type": "boolean"
        },
        "last": {
          "type": "boolean"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "numberOfElements": {
          "type": "integer",
          "format": "int32"
        },
        "size": {
          "type": "integer",
          "format": "int32"
        },
        "sort": {
          "$ref": "#/definitions/Sort"
        },
        "totalElements": {
          "type": "integer",
          "format": "int64"
        },
        "totalPages": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "ZoneRules": {
      "properties": {
        "fixedOffset": {
          "type": "boolean"
        },
        "transitionRules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ZoneOffsetTransitionRule"
          }
        },
        "transitions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ZoneOffsetTransition"
          }
        }
      }
    },
...
    "Discussion": {
      "properties": {
        "ask": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "registering": {
          "$ref": "#/definitions/ZonedDateTime"
        },
        "solutions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Solutions"
          }
        },
        "username": {
          "type": "string"
        }
      }
    },
...
    "Homework": {
      "properties": {
        "hwQuestions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/HwQuestion"
          }
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}